refactor: DiscordClient 재조정
