+++ Moyoy-Api/src/test/java/com/moyoy/api/user/application/UserServiceTest.java
@@ -27,8 +27,8 @@ import com.moyoy.domain.user.User;
 import com.moyoy.domain.user.UserCreate;
 import com.moyoy.domain.user.UserRepository;
 
-import com.moyoy.infra.database.mysql.common.UserRankingQueryRepository;
-import com.moyoy.infra.database.mysql.common.UserRankingView;
+import com.moyoy.infra.database.mysql.query.UserRankingQueryRepository;
+import com.moyoy.infra.database.mysql.query.UserRankingView;
 
 @ExtendWith(MockitoExtension.class)
 class UserServiceTest {
