+++ Moyoy-Infra/src/main/java/com/moyoy/infra/external/config/FeignClientConfig.java
@@ -1,4 +1,4 @@
-package com.moyoy.infra.external.support.config;
+package com.moyoy.infra.external.config;
 
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
