+++ Moyoy-Infra/src/main/java/com/moyoy/infra/external/discord/DiscordApi.java
@@ -1,4 +1,4 @@
-package com.moyoy.infra.external.discord.feign;
+package com.moyoy.infra.external.discord;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -6,10 +6,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import com.moyoy.infra.external.discord.dto.DiscordClientRequest;
-import com.moyoy.infra.external.support.config.DiscordFeignConfig;
+import com.moyoy.infra.external.discord.support.DiscordFeignConfig;
 
 @FeignClient(name = "discordWebhookClient", url = "https://discord.com/api/webhooks", configuration = DiscordFeignConfig.class)
-interface DiscordFeignClient {
+interface DiscordApi {
 
 	@PostMapping("/{id}/{token}")
 	void sendNotification(
