--- Moyoy-Infra/src/main/java/com/moyoy/infra/external/discord/feign/DiscordClient.java
@@ -1,8 +1,0 @@
-package com.moyoy.infra.external.discord.feign;
-
-import com.moyoy.infra.external.discord.dto.DiscordClientRequest;
-
-public interface DiscordClient {
-
-	void sendNotification(DiscordClientRequest discordClientRequest);
-}
