+++ Moyoy-Infra/src/main/java/com/moyoy/infra/external/github/repo/GithubRepoFeignClient.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import com.moyoy.infra.external.support.config.GithubFeignConfig;
+import com.moyoy.infra.external.github.support.GithubFeignConfig;
 
 import feign.Response;
